pekko: update disclaimer (#69)
I think this makes the relationship with the PMC clearer. I considered
mentioning that we actually are on the PMC as well, but decided against
it to err on the safe side of setting expectations.

I think "Apache Pekko will continue to be Free and Open Source Software" should now go without saying, but if you prefer keeping it I'm OK with that as well.

Trademark disclaimer following https://apache.org/foundation/marks/faq/#attribution

When this change is good to merge I'll also PR it to the campaign page.